now scraping for prices from amazon links in addition to stockx links

+++ models/pricelist.js
@@ -13,7 +13,7 @@ module.exports = function(sequelize, DataTypes) {
     name: {type: DataTypes.STRING,
         allowNull: false,
     },
-    price: {type: DataTypes.INTEGER,
+    price: {type: DataTypes.FLOAT,
         allowNull: false,
     },
     no_price: {type: DataTypes.BOOLEAN,
